+++ CHANGELOG.md
@@ -2,5 +2,5 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Percursus/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Percursus/releases)
 
-- Updated: deDE, enUS, ruRU localization
-- Minor code adjustments
+- Updated: Race times for the expansion 'Dragonflight'
+- Updated: ruRU localization (#2 - Thanks to Hollicsh)